Squashed 'rocksdb-sys/rocksdb/' changes from c38b075..2d5aaf5
9808491 Release 4.11.2
fbd1c98 Fix recovery for WALs without data for all CFs
c743289 Fix GetSortedWalFiles when log recycling enabled
2be165b Fix a crash when compaction fails to open a file
0f4c98d Increase version to 4.11.1
c75f4fa fix data race in NewIndexIterator() in block_based_table_reader.cc
56f497b Mitigate regression bug of options.max_successive_merges hit during DB Recovery
051c88d Compaction stats printing: "batch" => "commit group"
071893e Fix a destruction order issue in ThreadStatusUpdater
7d7234a Change HISTORY.md for release 4.11
d641bb8 read_options.background_purge_on_iterator_cleanup to cover forward iterator and log file closing too.
e98b76d A utility function to help users migrate DB after options change
4001e79 Ignore write stall triggers when auto-compaction is disabled
REVERT: c38b075 Update HISTORY.md
REVERT: 8f399e3 Update HISTORY.md
REVERT: 98d0b78 Added check_snapshot option in the DB's AddFile function (#1261)
REVERT: 9fd68b7 set travis open file descriptor limit
REVERT: 59ddb50 Fix travis build break
REVERT: f4d9863 Added SetOptions support to RocksJava (#1243)
REVERT: 7882cb9 Make DBOptionsTest::EnableAutoCompactionAndTriggerStall less falky
REVERT: 44f5cc5 Add time series database (resubmitted)
REVERT: 7c4615c A utility function to help users migrate DB after options change
REVERT: 5bb0a7f Update appveyor.yml
REVERT: 86396cc Update appveyor.yml
REVERT: c1db098 Update appveyor.yml
REVERT: 7da2eaf Update appveyor.yml
REVERT: 34723b4 Cleanup unused variable pending_fsync_.
REVERT: 7cc0dbd cat all logs in sandcastle output
REVERT: 9253767 Correct geHistogramData() -> getHistogramData() (#1257)
REVERT: f35b16f db_bench add an option of --base_background_compactions
REVERT: c3a4bea Fix flaky test `ObsoleteFiles`
REVERT: 8234faa Fix failed test
REVERT: 4990c0d Remove deprecated LEVELDB_PLATFORM_POSIX
REVERT: 7323e4c Fix clang on macOS
REVERT: ee027fc Ignore write stall triggers when auto-compaction is disabled
REVERT: e4609a7 Fix Windows build issues (#1253)
REVERT: 2306167 Fix clang build failure and refactor unit test
REVERT: 343304e Use StopWatch to do statistic job in db_impl_add_file.cc
REVERT: cdc4eb6 Add a GetComparator() function to the ColumnFamilyHandle base class so that the user's comparator can be retrieved.
REVERT: 712dd27 Build break fixes
REVERT: 0155c73 Fix parallel tests `make check -j`
REVERT: c49ea68 Fix to enable running CI jobs locally
REVERT: 726c2f7 Build break fix
REVERT: d51dc96 Experiments on column-aware encodings
REVERT: c116b47 Persistent Read Cache (part 6) Block Cache Tier Implementation
REVERT: 64046e5 Write a benchmark to emulate time series data

git-subtree-dir: rocksdb-sys/rocksdb
git-subtree-split: 2d5aaf572437d21eca206af737bf7b7b5a624c42

+++ db/db_impl.cc
@@ -716,6 +716,16 @@ uint64_t DBImpl::FindMinLogContainingOutstandingPrep() {
   return min_log;
 }
 
+void DBImpl::ScheduleBgLogWriterClose(JobContext* job_context) {
+  if (!job_context->logs_to_free.empty()) {
+    for (auto l : job_context->logs_to_free) {
+      AddToLogsToFreeQueue(l);
+    }
+    job_context->logs_to_free.clear();
+    SchedulePurge();
+  }
+}
+
 // * Returns the list of live files in 'sst_live'
 // If it's doing full scan:
 // * Returns the list of all files in the filesystem in
@@ -1566,10 +1576,16 @@ Status DBImpl::RecoverLogFiles(const std::vector<uint64_t>& log_numbers,
       // we just ignore the update.
       // That's why we set ignore missing column families to true
       bool has_valid_writes = false;
+      // If we pass DB through and options.max_successive_merges is hit
+      // during recovery, Get() will be issued which will try to acquire
+      // DB mutex and cause deadlock, as DB mutex is already held.
+      // The DB pointer is not needed unless 2PC is used.
+      // TODO(sdong) fix the allow_2pc case too.
       status = WriteBatchInternal::InsertInto(
           &batch, column_family_memtables_.get(), &flush_scheduler_, true,
-          log_number, this, false /* concurrent_memtable_writes */,
-          next_sequence, &has_valid_writes);
+          log_number, db_options_.allow_2pc ? this : nullptr,
+          false /* concurrent_memtable_writes */, next_sequence,
+          &has_valid_writes);
       // If it is the first log file and there is no column family updated
       // after replaying the file, this file may be a stale file. We ignore
       // sequence IDs from the file. Otherwise, if a newer stale log file that
@@ -1685,7 +1701,7 @@ Status DBImpl::RecoverLogFiles(const std::vector<uint64_t>& log_numbers,
       // recovered and should be ignored on next reincarnation.
       // Since we already recovered max_log_number, we want all logs
       // with numbers `<= max_log_number` (includes this one) to be ignored
-      if (flushed) {
+      if (flushed || cfd->mem()->GetFirstSequenceNumber() == 0) {
         edit->SetLogNumber(max_log_number + 1);
       }
       // we must mark the next log number as used, even though it's
@@ -2988,8 +3004,9 @@ void DBImpl::BGWorkCompaction(void* arg) {
 
 void DBImpl::BGWorkPurge(void* db) {
   IOSTATS_SET_THREAD_POOL_ID(Env::Priority::HIGH);
-  TEST_SYNC_POINT("DBImpl::BGWorkPurge");
+  TEST_SYNC_POINT("DBImpl::BGWorkPurge:start");
   reinterpret_cast<DBImpl*>(db)->BackgroundCallPurge();
+  TEST_SYNC_POINT("DBImpl::BGWorkPurge:end");
 }
 
 void DBImpl::UnscheduleCallback(void* arg) {
@@ -3004,20 +3021,32 @@ void DBImpl::UnscheduleCallback(void* arg) {
 void DBImpl::BackgroundCallPurge() {
   mutex_.Lock();
 
-  while (!purge_queue_.empty()) {
-    auto purge_file = purge_queue_.begin();
-    auto fname = purge_file->fname;
-    auto type = purge_file->type;
-    auto number = purge_file->number;
-    auto path_id = purge_file->path_id;
-    auto job_id = purge_file->job_id;
-    purge_queue_.pop_front();
+  // We use one single loop to clear both queues so that after existing the loop
+  // both queues are empty. This is stricter than what is needed, but can make
+  // it easier for us to reason the correctness.
+  while (!purge_queue_.empty() || !logs_to_free_queue_.empty()) {
+    if (!purge_queue_.empty()) {
+      auto purge_file = purge_queue_.begin();
+      auto fname = purge_file->fname;
+      auto type = purge_file->type;
+      auto number = purge_file->number;
+      auto path_id = purge_file->path_id;
+      auto job_id = purge_file->job_id;
+      purge_queue_.pop_front();
 
-    mutex_.Unlock();
-    Status file_deletion_status;
-    DeleteObsoleteFileImpl(file_deletion_status, job_id, fname, type, number,
-                           path_id);
-    mutex_.Lock();
+      mutex_.Unlock();
+      Status file_deletion_status;
+      DeleteObsoleteFileImpl(file_deletion_status, job_id, fname, type, number,
+                             path_id);
+      mutex_.Lock();
+    } else {
+      assert(!logs_to_free_queue_.empty());
+      log::Writer* log_writer = *(logs_to_free_queue_.begin());
+      logs_to_free_queue_.pop_front();
+      mutex_.Unlock();
+      delete log_writer;
+      mutex_.Lock();
+    }
   }
   bg_purge_scheduled_--;
 
@@ -3083,6 +3112,8 @@ void DBImpl::BackgroundCallFlush() {
   bool made_progress = false;
   JobContext job_context(next_job_id_.fetch_add(1), true);
   assert(bg_flush_scheduled_);
+
+  TEST_SYNC_POINT("DBImpl::BackgroundCallFlush:start");
 
   LogBuffer log_buffer(InfoLogLevel::INFO_LEVEL, db_options_.info_log.get());
   {
@@ -3655,6 +3686,9 @@ static void CleanupIteratorState(void* arg1, void* arg2) {
     state->mu->Lock();
     state->super_version->Cleanup();
     state->db->FindObsoleteFiles(&job_context, false, true);
+    if (state->background_purge) {
+      state->db->ScheduleBgLogWriterClose(&job_context);
+    }
     state->mu->Unlock();
 
     delete state->super_version;
